requirements.txt
azure-functions==1.21.3
fastapi==0.115.6
azure-cosmos==4.9.0
openai==1.82.0
pydantic==2.10.4
uvicorn==0.34.0
aiohttp==3.11.14
